# projects/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/register": {
      "post": {
        "description": "校验验证码后创建用户并返回 JWT（1h）。须先 GET /auth/verificationCode?type=register&key={username}",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "用户注册",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "description": "Local 策略校验账号 + 验证码，返回 JWT（2h）。须先 GET /auth/verificationCode?type=login&key={username}",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "用户名密码登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/verificationCode": {
      "get": {
        "description": "Query：type=login|register，key=用户名。开发环境响应含明文 code",
        "operationId": "AuthController_verificationCode",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "query",
            "description": "用户名，与 type 组成 Redis 键",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "description": "验证码场景",
            "schema": {
              "type": "string",
              "enum": [
                "login",
                "register"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取验证码",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/upload": {
      "post": {
        "description": "multipart 字段 file；允许 .jpg .jpeg .png .md",
        "operationId": "AuthController_upload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "上传文件",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/hydrology/collection": {
      "get": {
        "description": "可按 cityName、sttp 筛选；sttp 支持多值 OR",
        "operationId": "HydrologyController_collection",
        "parameters": [
          {
            "name": "cityName",
            "required": false,
            "in": "query",
            "description": "城市名筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sttp",
            "required": false,
            "in": "query",
            "description": "站类，可多选 ?sttp=PP&sttp=ZQ 或 ?sttp=PP,ZQ",
            "schema": {
              "$ref": "#/components/schemas/Object"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "水文站列表",
        "tags": [
          "hydrology"
        ]
      }
    },
    "/v1/hydrology/radar": {
      "get": {
        "description": "距经纬度 2km 内站点",
        "operationId": "HydrologyController_radar",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "纬度 -90～90",
            "schema": {
              "example": 31.2,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "经度 -180～180",
            "schema": {
              "example": 121.5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "雷达范围水文站",
        "tags": [
          "hydrology"
        ]
      }
    },
    "/v1/hydrology/radarDetail": {
      "get": {
        "description": "2km 内站点及各站最新 stDetail 记录",
        "operationId": "HydrologyController_radarDetail",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "纬度 -90～90",
            "schema": {
              "example": 31.2,
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "经度 -180～180",
            "schema": {
              "example": 121.5,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "雷达详情",
        "tags": [
          "hydrology"
        ]
      }
    },
    "/v1/tidal/data": {
      "get": {
        "description": "站点 code + 日期",
        "operationId": "TidalController_data",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "query",
            "description": "港口/站点 code",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "description": "日期 YYYY-MM-DD",
            "schema": {
              "example": "2026-06-04",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取潮汐数据",
        "tags": [
          "tidal"
        ]
      }
    },
    "/v1/tidal/areas": {
      "get": {
        "operationId": "TidalController_areas",
        "parameters": [
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "国家代码，如 CN",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取区域列表",
        "tags": [
          "tidal"
        ]
      }
    },
    "/v1/tidal/point": {
      "get": {
        "description": "按 parentId（tide_category）查询港口列表",
        "operationId": "TidalController_point",
        "parameters": [
          {
            "name": "parentId",
            "required": true,
            "in": "query",
            "description": "区域分类 parentId（tide_category.id）",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "日期 YYYY-MM-DD，默认当天",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": false,
            "in": "query",
            "description": "生活指数天数 1～7，默认 3",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "获取港口标记点",
        "tags": [
          "tidal"
        ]
      }
    },
    "/v1/weather/minutely/10m": {
      "get": {
        "description": "和风 locationId",
        "operationId": "WeatherController_minutely10m",
        "parameters": [
          {
            "name": "location",
            "required": true,
            "in": "query",
            "description": "和风 locationId",
            "schema": {
              "example": "101010100",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "10 分钟降雨量",
        "tags": [
          "weather"
        ]
      }
    },
    "/v1/weather/airquality/daily/{latitude}/{longitude}": {
      "get": {
        "operationId": "WeatherController_airqualityDaily",
        "parameters": [
          {
            "name": "latitude",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "longitude",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "空气质量（日）",
        "tags": [
          "weather"
        ]
      }
    },
    "/v1/weather/basic/one": {
      "post": {
        "description": "24h 天气、生活指数、日出日落",
        "operationId": "WeatherController_basicOne",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherBasicSingleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "基础天气（单点）",
        "tags": [
          "weather"
        ]
      }
    },
    "/v1/weather/basic": {
      "post": {
        "description": "按 location 数组顺序逐项拉取",
        "operationId": "WeatherController_basic",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeatherBasicBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "基础天气（批量）",
        "tags": [
          "weather"
        ]
      }
    },
    "/v1/wechat/auth/login": {
      "post": {
        "description": "wx.login 的 code 换 openid 并签发 JWT。可选 phoneCode（getPhoneNumber）、avatarUrl",
        "operationId": "WechatController_authLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatAuthLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "小程序授权登录（推荐）",
        "tags": [
          "wechat"
        ]
      }
    },
    "/v1/wechat/auth/me": {
      "get": {
        "description": "需 Bearer JWT",
        "operationId": "WechatController_authMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "当前登录用户",
        "tags": [
          "wechat"
        ]
      }
    },
    "/v1/wechat/login": {
      "post": {
        "description": "仅传 code，等价于 auth/login 的精简响应",
        "operationId": "WechatController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatCodeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "小程序登录（兼容）",
        "tags": [
          "wechat"
        ]
      }
    },
    "/v1/wechat/location": {
      "post": {
        "description": "WGS84 坐标转地址，需配置 TENCENT_LBS_KEY 或 QQMAP_KEY",
        "operationId": "WechatController_location",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "逆地理编码",
        "tags": [
          "wechat"
        ]
      }
    },
    "/v1/test/test": {
      "get": {
        "operationId": "TestController_test",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "连通性测试",
        "tags": [
          "test"
        ]
      }
    },
    "/v1/test/restricted": {
      "get": {
        "description": "同一 IP 当日匿名访问超过 API_GUEST_RESTRICTED_MAX（默认 3）后须 Bearer JWT",
        "operationId": "TestController_restricted",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "限流测试接口",
        "tags": [
          "test"
        ]
      }
    }
  },
  "info": {
    "title": "Tidal API",
    "description": "Tidal 业务 API（OpenAPI 3，由 @nestjs/swagger 生成）",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "登录名，唯一"
          },
          "password": {
            "type": "string",
            "description": "明文密码（服务端 MD5 入库）"
          },
          "code": {
            "type": "string",
            "description": "须先 GET /auth/verificationCode 获取"
          },
          "cover": {
            "type": "string",
            "description": "头像 URL"
          }
        },
        "required": [
          "username",
          "password",
          "code"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "须先 GET /auth/verificationCode 获取"
          }
        },
        "required": [
          "username",
          "password",
          "code"
        ]
      },
      "Object": {
        "type": "object",
        "properties": {}
      },
      "WeatherBasicSingleDto": {
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "example": "101010100"
          }
        },
        "required": [
          "location"
        ]
      },
      "WeatherBasicBatchDto": {
        "type": "object",
        "properties": {
          "location": {
            "example": [
              "101010100",
              "101020100"
            ],
            "description": "和风 locationId 数组",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "location"
        ]
      },
      "WechatAuthLoginDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "wx.login() 返回的 code，必填"
          },
          "phoneCode": {
            "type": "string",
            "description": "getPhoneNumber 组件返回的 code"
          },
          "avatarUrl": {
            "type": "string",
            "description": "用户头像 URL，写入 user.cover"
          }
        },
        "required": [
          "code"
        ]
      },
      "WechatCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "wx.login() 返回的临时 code"
          }
        },
        "required": [
          "code"
        ]
      },
      "WechatLocationDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "example": 39.9,
            "description": "纬度 WGS84"
          },
          "longitude": {
            "type": "number",
            "example": 116.4,
            "description": "经度 WGS84"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      }
    }
  }
}